% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{appendCsvTable}
\alias{appendCsvTable}
\title{Append a table to a csv file}
\usage{
appendCsvTable(table, filePath)
}
\arguments{
\item{table}{A dataframe to write to the csv}

\item{filePath}{The path to the csv file}
}
\value{
invisible file path
}
\description{
This function appends a data frame to an existing csv file or creates a new
file if it does not yet exist.
}
