% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plpDataBuilders.R
\name{buildAndExecuteCohort}
\alias{buildAndExecuteCohort}
\title{Build and execute a cohort definition from JSON}
\usage{
buildAndExecuteCohort(
  connection,
  jsonPath,
  cohortId,
  cdmSchema,
  vocabularySchema,
  cohortSchema,
  cohortTable,
  targetDialect = "postgresql"
)
}
\arguments{
\item{connection}{A \code{DatabaseConnector} connection object}

\item{jsonPath}{The path to the cohort-definition JSON file}

\item{cohortId}{The cohort definition id to assign in the target cohort table}

\item{cdmSchema}{The CDM database schema}

\item{vocabularySchema}{The vocabulary database schema}

\item{cohortSchema}{The schema where the cohort table will be written}

\item{cohortTable}{The name of the target cohort table}

\item{targetDialect}{The SQL dialect used by the target database}
}
\value{
Invisibly returns the result of \code{DatabaseConnector::executeSql()}
}
\description{
This function reads a cohort definition from a JSON file, converts it to a CirceR
cohort expression, renders and translates the corresponding SQL for the target database, and executes it using the supplied connection
}
