% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputationMlHelpers.R
\name{buildMissingIndicatorAnalysisRef}
\alias{buildMissingIndicatorAnalysisRef}
\title{Build missing indicator analysis reference rows}
\usage{
buildMissingIndicatorAnalysisRef(
  analysisRef,
  sourceCovariateIds,
  sourceInfo,
  indicatorAnalysisIds
)
}
\arguments{
\item{analysisRef}{The original analysis reference table}

\item{sourceCovariateIds}{The source covariate identifiers}

\item{sourceInfo}{A table containing metadata for the source covariates}

\item{indicatorAnalysisIds}{The new indicator analysis identifiers}
}
\value{
An analysis reference table for the new indicator covariates
}
\description{
This function creates a new \code{analysisRef} rows for missingness indicator
covariates derived from imputable covariates
}
