% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputationMlHelpers.R
\name{buildMissingIndicatorCovariateRef}
\alias{buildMissingIndicatorCovariateRef}
\title{Build missing indicator covariate reference rows}
\usage{
buildMissingIndicatorCovariateRef(
  covariateRef,
  sourceCovariateIds,
  indicatorCovariateIds,
  indicatorAnalysisIds
)
}
\arguments{
\item{covariateRef}{The original covariate reference table}

\item{sourceCovariateIds}{The source covariate identifiers}

\item{indicatorCovariateIds}{The new indicator covariate identifiers}

\item{indicatorAnalysisIds}{The new indicator analysis covariates}
}
\value{
A covariate reference table for the new indicator covariates
}
\description{
This function creates new \code{covariateRef} rows for missingness indicator
covaraites derived from imputatble source covariates
}
