% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingnessPatterns.R
\name{buildPatternMissingnessMatrix}
\alias{buildPatternMissingnessMatrix}
\title{Build a pattern missingness matrix}
\usage{
buildPatternMissingnessMatrix(patterns, covariateIds)
}
\arguments{
\item{patterns}{A list of missingness patterns}

\item{covariateIds}{The covariate ids to represent in the matrix}
}
\value{
A binary matrix with covariates in rows and patterns in columns
}
\description{
This function converts a list of pattern specifications into a binary matrix
indicating which covariates are set missing by each pattern
}
