% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plpDataBuilders.R
\name{buildPopulationPLPData}
\alias{buildPopulationPLPData}
\title{Build a PLP data object for a subset population}
\usage{
buildPopulationPLPData(
  selectedRowIds,
  population,
  covariateValues,
  covariateRef,
  analysisRef,
  measurementConceptIds = integer(),
  templatePLPData,
  includeFolds = TRUE
)
}
\arguments{
\item{population}{A population table}

\item{covariateValues}{A covariate-values table}

\item{covariateRef}{A covariate reference table}

\item{analysisRef}{An analysis reference table}

\item{measurementConceptIds}{Optional measurement concept ids to retain}

\item{templatePLPData}{A template \code{plpData} object whose structure and metadata should
be preserved}

\item{includeFolds}{Logical variable indicating whether default folds table should added
to the output}

\item{selectedRowids}{The row ids to include in the object}
}
\value{
A list containing the rebuilt \code{plpData} object and the filtered population table
}
\description{
This function constructs a \code{plpData} object for a selected set of row ids
by subsetting the popultaion, covariates and reference tables from a larger PLP data source
}
