% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgboostImputer.R
\name{buildXgboostArguments}
\alias{buildXgboostArguments}
\title{Build iterative xgboost imputation training argumnets}
\usage{
buildXgboostArguments(settings, isBinaryTarget)
}
\arguments{
\item{settings}{The iterative xgboost imputer settings}

\item{isBinaryTarget}{Logical variable indicating whether the current target covariate is binary}
}
\value{
A named list of xgboost training parameters
}
\description{
This function constructs the xgboost parameter list used to fit one imputation model,
adapting the objective and evaluation metric to the target covariate type
}
