% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgboostImputer.R
\name{buildXgboostDesignMatrix}
\alias{buildXgboostDesignMatrix}
\title{Build a design matrix for the iterative xgboost imputation}
\usage{
buildXgboostDesignMatrix(state, rowIndex, targetIndex)
}
\arguments{
\item{state}{An iterative xgboost imputation training or apply state object}

\item{targetIndex}{The column index of the target covariate currently being modeled}

\item{rowIdIndex}{The row idices to include in the design matrix}
}
\value{
A sparse design matrix, or \code{NULL} when no predictors are available
}
\description{
This function constructs the predictor matrix used to fit or apply one iterative
xgboost imputation model. It combines sparse zero-default predictors with the current filled values of the other
imputable dense target covariates
}
