% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingnessHelpers.R
\name{calcScore}
\alias{calcScore}
\title{Compute a missingness score}
\usage{
calcScore(
  covariates,
  rowIds,
  causeCovariateIds,
  weights = NULL,
  standardize = TRUE
)
}
\arguments{
\item{covariates}{A long format covariate table}

\item{rowIds}{The row identifiers to score}

\item{causeCovariateIds}{The covariate identifiers used to build the score}

\item{weights}{Optional weights for the cause covariates}

\item{standardize}{A logical paramter indicating whether the score should be
standardized}
}
\value{
A data frame or vector containing scores for each individual
}
\description{
This function computes a score for each individual used to drive missingness assignment from one or more
cause variables
}
