% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{collapseIds}
\alias{collapseIds}
\title{Collapse identifiers into one label}
\usage{
collapseIds(x)
}
\arguments{
\item{x}{A vector of identifiers}
}
\value{
A single character string containing the sorted unique identifiers, or NA_character_ when \code{x} is \code{NULL}
}
\description{
This function converts a vector of identifiers into a single label separated by underscores
}
