% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageHelpers.R
\name{collectIfNeeded}
\alias{collectIfNeeded}
\title{Materialize an object when possible}
\usage{
collectIfNeeded(x)
}
\arguments{
\item{x}{An object that may require collection into memory}
}
\value{
The collected object when collection is supported, otherwise the original input
}
\description{
This function calls \code{dplyr::collect()} on database objects when possible, otherwise
it returns the input unchanged
}
