% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputationMlHelpers.R
\name{computeBinaryMode}
\alias{computeBinaryMode}
\title{Compute the binary mode}
\usage{
computeBinaryMode(x)
}
\arguments{
\item{x}{A vector of observed binary covariate values}
}
\value{
The modal binary value
}
\description{
This function returns the mode value of a binary vector
}
