% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputationMethods.R
\name{createPLPImputationMethods}
\alias{createPLPImputationMethods}
\title{Create default PLP imputation methods}
\usage{
createPLPImputationMethods(
  includeMissingIndicator = c(FALSE, TRUE),
  sklearnRandomState = 432,
  missForestSeed = 432L,
  iterativeXGBoostSeed = 432L,
  pmmTrainSeed = 432L,
  pmmTestSeed = 432L
)
}
\arguments{
\item{includeMissingIndicator}{Logical vector indicating whether to create imputationmethods
with or without missingness  indicators or both}

\item{sklearnRandomState}{The random state used fot the sklearn iterative imputer when available}

\item{missForestSeed}{The seed used for missForest}

\item{iterativeXGBoostSeed}{The seed used for iterative XGBoost imputation}

\item{pmmTrainSeed}{The seed used when fitting PMM on the train data}

\item{pmmTestSeed}{The seed used when applying PMM to the test data}
}
\value{
A list of \code{featureEngineeringSettings} objects
}
\description{
This function builds a list of default imputation method settings used in the
simulation workflow. It creates variants with and without missingness indicators and checks
whether the required packages are installed
}
