% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{createPLPPredictionModels}
\alias{createPLPPredictionModels}
\title{Set the PLP prediction models}
\usage{
createPLPPredictionModels(
  lassoSeed = 12L,
  xgboostSeed = 12L,
  transformerSeed = 12L,
  transformerDevice = NULL
)
}
\arguments{
\item{lassoSeed}{A random seed for the lasso prediction model}

\item{xgboostSeed}{A random seed for the XGBoost prediction model}

\item{transformerSeed}{A random seed for the transformer model}

\item{transformerDevice}{A parameter to specify the device to run the transformer model on}
}
\value{
A named list containing three PLP model specifications
}
\description{
This function creates named PLP prediction models required for the
simulation study, using default settings.
}
