% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{defaultSimulationInputs}
\alias{defaultSimulationInputs}
\title{Determine the default simulation covariates}
\usage{
defaultSimulationInputs(
  targetCovariateId,
  causeCovariateIds = NULL,
  targetCovariateIdPerMech = NULL,
  causeCovariateIdsPerMech = NULL,
  patterns = NULL
)
}
\arguments{
\item{targetCovariateId}{The default target covariate identifier (or a vecotr of identifiers)}

\item{causeCovariateIds}{Optional default cause covariate identifiers}

\item{targetCovariateIdPerMech}{Optional mechanism-specific target covariate identifiers}

\item{causeCovariateIdsPerMech}{Optional mechanism-specific cause covariate identifiers}

\item{patterns}{Optional list of explicit pattern specificiations}
}
\value{
A vector of required covariate identifiers
}
\description{
This function collects the target and cause covariate identifiers that must be
present in the complete input data. It supports mechanism-specific inputs and
multivariate pattern specifications
}
