% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{evaluationStatsHelper}
\alias{evaluationStatsHelper}
\title{Parse evaluation table values}
\usage{
evaluationStatsHelper(table)
}
\arguments{
\item{table}{A data frame containing the evaluation results}
}
\value{
A data frame with parsed evaluation values for aggregation
}
\description{
This function parses evaluation metrics into a consistent format such that they can be summarised
across simulation runs in a later stage.
}
