% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgboostImputer.R
\name{fitSingleXgboostModel}
\alias{fitSingleXgboostModel}
\title{Fit one iterative xgboost imputation model}
\usage{
fitSingleXgboostModel(state, targetCovariateId, settings, fitModelOnly = FALSE)
}
\arguments{
\item{state}{An iterative xgboost training state object}

\item{targetCovariateId}{The covariate id currently being imputed}

\item{settings}{The iterative xboost imputer settings}

\item{fitModelOnly}{Logical variable indicating whether only the fitted model should
be returned without generating predictions for missing values}
}
\value{
A list with elements \code{model} and \code{predictions}
}
\description{
This function fits the iterative xgboost imputation model for one covariate and, unless
\code{fitModelOnly = TRUE}, also predicts the missing value for that target
}
