% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgboostImputer.R
\name{fitXgboostImputer}
\alias{fitXgboostImputer}
\title{Fit the iterative xgboost imputer}
\usage{
fitXgboostImputer(prepared, settings)
}
\arguments{
\item{prepared}{A prepared PLP imputation object}

\item{settings}{The iterative xgboost imputer settings}
}
\value{
A fitted iterative xgboost imputer object
}
\description{
This function fits the full iterative xgboost imputer on prepared
training data. It selects the target covariates to impute, initializes the missing values,
performs the iterative update loop, and then fits the final per-covariate models during application
}
