% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputationMain.R
\name{formatImputationElapsed}
\alias{formatImputationElapsed}
\title{Format an imputation runtime}
\usage{
formatImputationElapsed(startTime, endTime = Sys.time())
}
\arguments{
\item{startTime}{The start time}

\item{endTime}{The end time. Defaults to the current time}
}
\value{
A character string describing the elapsed time in seconds
}
\description{
This function formats the elapsed time between two timestamp as a readable string in seconds
}
