% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputationHelpers.R
\name{getPLPImputationMetaKey}
\alias{getPLPImputationMetaKey}
\title{Map an imputation function to its metadata key}
\usage{
getPLPImputationMetaKey(funName)
}
\arguments{
\item{funName}{The name of the imputation function}
}
\value{
A string containing the metadata key for the imputer
}
\description{
This function returns the feature-engineering metadata key to store
fitted settings for a given PLP imputation function
}
