% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputationHelpers.R
\name{getPLPImputer}
\alias{getPLPImputer}
\title{Resolve a PLP imputation function}
\usage{
getPLPImputer(funName)
}
\arguments{
\item{funName}{The name of the imputation function to resolve}
}
\value{
A function with arguments \code{trainData}, \code{featureEngineeringSettings} and \code{done}
}
\description{
This function looks up an imputation function by name and returns a wrapper
that copies the PLP data object before execution. It supports both functions provided
by the PatientLevelPrediction package and local imputation functions defined in this package
}
