% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputationMethods.R
\name{getPmmExecutionSeed}
\alias{getPmmExecutionSeed}
\title{Get PMM execution seed}
\usage{
getPmmExecutionSeed(imputerSettings, done = FALSE)
}
\arguments{
\item{done}{Logical parameter indicating whether the imputer is being applied to new data using fitted settings}
}
\value{
A seed for PMM execution (or NULL for any other imputation method)
}
\description{
This function returns the PMM seed used for training or test set imputation. Non-PMM imputers return NULL
@param imputerSettings An imputer settings object
}
