% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgboostImputer.R
\name{implementXgboostImputer}
\alias{implementXgboostImputer}
\title{Implement iterative xgboost imputation to work with PLP}
\usage{
implementXgboostImputer(trainData, featureEngineeringSettings, done = FALSE)
}
\arguments{
\item{trainData}{A PLP training data object}

\item{featureEngineeringSettings}{Iterative XGBoost imputation sttings}
}
\description{
This function imputes the PLP data object using iterative imputation
}
