% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputationMlHelpers.R
\name{initializeLongImputations}
\alias{initializeLongImputations}
\title{Initialize long format imputations}
\usage{
initializeLongImputations(missingIndex, initialFill)
}
\arguments{
\item{missingIndex}{A long format table identifying missing row/covariate pairs}

\item{initialFill}{A list of initial fill values by covariate}
}
\value{
A long format table containing the initialized imputed values
}
\description{
This function creates the initial long format tale of imputed values for
missing cells using the supplied per covariate fill values
}
