% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputationMlHelpers.R
\name{isBinaryCovariate}
\alias{isBinaryCovariate}
\title{Check whether a covariate is binary}
\usage{
isBinaryCovariate(x)
}
\arguments{
\item{x}{A vector of observed covariate values}
}
\value{
A logical value indicating whether the values are binary
}
\description{
This function determines whether the observed values for a covariate are binary
}
