% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputationMethods.R
\name{isPmmImputerSettings}
\alias{isPmmImputerSettings}
\title{Check whether settings correspond to PMM imputer}
\usage{
isPmmImputerSettings(imputerSettings)
}
\arguments{
\item{imputerSettings}{An imputer settings object}
}
\value{
Logical indicating whether the imputer method is \code{PMM}
}
\description{
This function identifies whether an imputer settings object uses PMM
}
