% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plpDataBuilders.R
\name{keepMeasurements}
\alias{keepMeasurements}
\title{Retain selected measurement covariates}
\usage{
keepMeasurements(
  covariateRef,
  covariateValues,
  analysisRef,
  measurementConceptIds
)
}
\arguments{
\item{covariateRef}{A covariate reference table}

\item{covariateValues}{A covariates values table}

\item{analysisRef}{An analysis reference table}

\item{measurementConceptIds}{The measurement concept ids to retain}
}
\value{
A list containing the filtered \code{covariateRef}, \code{covariateValues}, \code{analysisRef}, and
the retained \code{measurementRef}
}
\description{
This function filters the measurement covariates to a requested st of measurement concept
ids while leaving other types of covariates untouched
}
