% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{loadRawEvaluationTables}
\alias{loadRawEvaluationTables}
\title{Load raw evaluation tables}
\usage{
loadRawEvaluationTables(folder)
}
\arguments{
\item{folder}{The simulation output folder containing the \code{raw} subfolder}
}
\value{
A list containing the raw evaluatin tables:
\code{evaluationTables}, \code{calibrationSummary}, \code{thresholdSummary},\code{demographicsSummary},
and \code{predictionDistribution}
}
\description{
This function reads the raw evaluation tables. Missing files are returned as \code{NULL}
}
