% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingnessHelpers.R
\name{logitInverse}
\alias{logitInverse}
\title{Compute the inverse logit}
\usage{
logitInverse(x)
}
\arguments{
\item{x}{A numeric value or vector}
}
\value{
A numeric vector of probabilities
}
\description{
This function maps numeric values to probabilities on a (0,1) scale using the
inverse logit transformation
}
