% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputationMlHelpers.R
\name{normalizeBinaryFlag}
\alias{normalizeBinaryFlag}
\title{Normalize flag values}
\usage{
normalizeBinaryFlag(x, default = "N")
}
\arguments{
\item{x}{A vector containing flag values}

\item{default}{The default value used when input values are missing or unrecognized}
}
\value{
A character vector containing normalized flag values
}
\description{
This function standardizes flag values to "Y" and "N" values
}
