% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputationMlHelpers.R
\name{normalizeIdKey}
\alias{normalizeIdKey}
\title{Normalize identifier keys}
\usage{
normalizeIdKey(x)
}
\arguments{
\item{x}{A vector of identifiers}
}
\value{
A vector of normalized identifier keys
}
\description{
This function converts identifiers to a consistent character key for matching
and look up
}
