% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingnessPatterns.R
\name{normalizeMissingFractionTargets}
\alias{normalizeMissingFractionTargets}
\title{Normalize target missing fractions}
\usage{
normalizeMissingFractionTargets(missingFractions, covariateIds)
}
\arguments{
\item{missingFractions}{A named vector of missingness fractions}

\item{covariateIds}{The covariate identifiers to align to}
}
\value{
A numeric vector of missingness fractions
}
\description{
This function standardizes the requested covariate level missing fractions into a names numeric vector
aligned with the supplied covariate identifiers
}
