% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingnessPatterns.R
\name{normalizePatternFractionTargets}
\alias{normalizePatternFractionTargets}
\title{Normalize pattern fractions}
\usage{
normalizePatternFractionTargets(patternFractions, nrOfPatterns)
}
\arguments{
\item{patternFractions}{A named vector of pattern fractions}

\item{nrOfPatterns}{The number of missingness patterns}
}
\value{
A numeric vector of pattern fractions
}
\description{
This function standardizes the requested pattern-frequency vector into a numeric vector aligned with
the supplied number of patterns
}
