% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingnessHelpers.R
\name{observedCovariateCheck}
\alias{observedCovariateCheck}
\title{Check that target covariates are observed}
\usage{
observedCovariateCheck(covariatesDf, targetCovariateIds)
}
\arguments{
\item{covariatesDf}{A long format covariate table}

\item{targetCovariateIds}{The covariate ids to check}
}
\value{
\code{TRUE} if all target covariates are observed, otherwise it throws an error
}
\description{
This function checks whether the target covariates are complete before simulating
missingness
}
