% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plpDataBuilders.R
\name{populationSubset}
\alias{populationSubset}
\title{Build a subset of the population given selected row ids}
\usage{
populationSubset(population, selectedRowIds)
}
\arguments{
\item{population}{A population table}

\item{selectedRowIds}{The row ids to retain}
}
\value{
A filtered population table
}
\description{
This helper filters PLP population tbale to the requested row ids while
preserving the original metadata
}
