% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgboostImputer.R
\name{predictSingleXgboostModel}
\alias{predictSingleXgboostModel}
\title{Predict from one iterative xgboost model}
\usage{
predictSingleXgboostModel(modelObject, newdata, rowCount = NULL)
}
\arguments{
\item{modelObject}{A fitted model object returned by \code{fitSingleXgboostModel()}}

\item{rowCount}{The number of rows to predict. Used when \code{newData} is \code{NULL} or  when
constant predictions are returned}

\item{newData}{A predictor matrix for the rows to be imputed}
}
\value{
A numeric vector of predicted values, or \code{NULL} when no model is available
}
\description{
This functin generates predictions for one fitted iterative xgboost imputation
model, handling both fitted xgboost models and constant fallback models
}
