% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgboostImputer.R
\name{prepareXgboostApplyState}
\alias{prepareXgboostApplyState}
\title{Prepare the iterative xboost apply state}
\usage{
prepareXgboostApplyState(prepared, fittedImputer)
}
\arguments{
\item{prepared}{A prepared PLP imputation object}

\item{fittedImputer}{Object with fitted iterative xgboost imputation settings}
}
\value{
A list containing the prepared apply state for iterative xgboost imputation
}
\description{
This helper prepares matrices, indices and storage objects needed
to apply a fitted iterative xgboost imputer to new data
}
