% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgboostImputer.R
\name{prepareXgboostTrainingState}
\alias{prepareXgboostTrainingState}
\title{Prepare the iterative xgboost imputation training state}
\usage{
prepareXgboostTrainingState(prepared, settings)
}
\arguments{
\item{prepared}{A prepared PLP imputation object}

\item{settings}{The iterative xgboost imputer settings}
}
\value{
A list containing the prepared training state for iterative xgboost imputation
}
\description{
This helper prepares the matrices, indices and storing objects
tneeded to fit the iterative xgboost needed to fit the iterative xgboost imputer on the training data
}
