% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingnessHelpers.R
\name{reportMissingFraction}
\alias{reportMissingFraction}
\title{Report the change in missingenss}
\usage{
reportMissingFraction(beforePLPData, afterPLPData, covariateIds = NULL)
}
\arguments{
\item{beforePLPData}{The original PLP data object}

\item{afterPLPData}{The PLP data object after missingness has been simulated}

\item{covariateIds}{Optional covariate ids to compare}
}
\value{
A data frame describing the change in missingness per covariate
}
\description{
This function compares the missingness before and after simulation
}
