% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputationMain.R
\name{runImputerWithOptionalSeed}
\alias{runImputerWithOptionalSeed}
\title{Run an imputer with an optional PMM seed}
\usage{
runImputerWithOptionalSeed(
  imputerFun,
  trainData,
  featureEngineeringSettings,
  done = FALSE
)
}
\arguments{
\item{imputerFun}{The imputation function to execute}

\item{trainData}{The PLP data object to impute}

\item{featureEngineeringSettings}{The imputer settings passed to the imputation function}

\item{done}{Logical parameter to indicate whether the imputer is being applied to
new data using the fitted settings}
}
\value{
The results returned by imputerFun
}
\description{
This function runs an imputation function and appllies a seed only when the imputer settings correspond
to PMM
}
