% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{runPlpModels}
\alias{runPlpModels}
\title{Fit and evaluate PLP models across imputed data sets}
\usage{
runPlpModels(
  imputationResults,
  modelOverview = NULL,
  preprocessSettings = PatientLevelPrediction::createPreprocessSettings(minFraction =
    0.001, normalize = TRUE, removeRedundancy = TRUE),
  hyperparameterSettings = PatientLevelPrediction::createHyperparameterSettings(),
  analysisPrefix = "sim",
  analysisPath = tempdir()
)
}
\arguments{
\item{imputationResults}{Named list of imputation results. Each element contains a \code{trainImputed} object and
optionally a \code{testImputed} object}

\item{modelOverview}{A named list of PLP model specifications.
If \code{NULL} models are created with \code{createPLPPredictionModels()}}

\item{preprocessSettings}{Preprocessing setings. By default, covariates occurring in less than 0.1 percent of the
population are removed, normalization is enabled and redundant covariates are removed}

\item{hyperparameterSettings}{Hyperparameter search settings}

\item{analysisPrefix}{Prefix used when constructing \code{analysisId} supplied to \code{\link[PatientLevelPrediction:fitPlp]{fitPlp()}}}

\item{analysisPath}{Directory in which PLP analysis object are written}
}
\value{
A list with model results for each imputation/prediction combination, recording the fitted model,
test predictions, evaluation object, success status, and any error message
}
\description{
Fits multiple PLP models for each imputed data set, generates predictions for
training and test set, evaluates model performance, and collects both detailed model results and evaluation metrics
}
