% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{saveSummarisedResults}
\alias{saveSummarisedResults}
\title{Save summarised evaluation results}
\usage{
saveSummarisedResults(aggregatedTables, folder)
}
\arguments{
\item{aggregatedTables}{A list of aggregated evaluation tables}

\item{folder}{The output folder where the summarised results should be saved}
}
\value{
Invisibly returns \code{NULL}
}
\description{
This function writes the aggregated evaluation tables to disk. It stores
one master csv per table type and also writes scenario specific subsets
}
