% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingnessHelpers.R
\name{softmax}
\alias{softmax}
\title{Compute softmax probabilities}
\usage{
softmax(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A numeric vector of softmax probabilities
}
\description{
This function converts a numeric score vector into normalized probabilities
that sum to one
}
