% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputationUtils.R
\name{subsetPlpDataRows}
\alias{subsetPlpDataRows}
\title{Subset a PLP data object by row identifiers}
\usage{
subsetPlpDataRows(plpData, rowIds)
}
\arguments{
\item{plpData}{A PLP data object}

\item{rowIds}{The row identifiers to retain}
}
\value{
A PLP data object restricted to the requested rows
}
\description{
This function returns a PLP data object containing only the
requested row identfiers while preserving the original associated labels,
folds and covariate reference tables
}
