% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{summariseModelRunStatus}
\alias{summariseModelRunStatus}
\title{Summarize PLP model run status}
\usage{
summariseModelRunStatus(modelResults)
}
\arguments{
\item{modelResults}{Result object returned by \code{runPlpModels()}}
}
\value{
A list containing the completion status and the reported error messages
}
\description{
Examines the output of \code{runPlpModels()} and determines whether the requested model runs completed
successfully, partially failed (e.g. for one of the imputation methods/mechanisms/ratios) or all failed
}
