% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{summariseOverSims}
\alias{summariseOverSims}
\title{Summarise evaluation results across simulation runs}
\usage{
summariseOverSims(table)
}
\arguments{
\item{table}{A raw evaluation table containing evaluation statistics of multiple simulation runs}
}
\value{
A data frame summarising numeric results across simulations
}
\description{
This function aggregates an evaluation table across simulation runs by
grouping on the scenario and result identifiers and computing mean and standard deviation for the numeric results columns
}
