% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingnessHelpers.R
\name{transformScore}
\alias{transformScore}
\title{Transform a missingness score}
\usage{
transformScore(score, type = "RIGHT")
}
\arguments{
\item{score}{A numeric score vector}

\item{type}{The score transformation type: \code{RIGHT}, \code{LEFT}, \code{MID}, \code{TAIL}}
}
\value{
A transformed numeric score vector
}
\description{
This function transforms a row level score so different missingness shapes can be
created, such as higher missingness in the right tail, left tail, middle or
extremes of the distribution
}
