% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputationMlHelpers.R
\name{updateImputedValues}
\alias{updateImputedValues}
\title{Update imputed values for one target covariate}
\usage{
updateImputedValues(currentImputedValues, targetCovariateId, predictions)
}
\arguments{
\item{currentImputedValues}{The current long format imputed value table}

\item{targetCovariateId}{The covariate identifier being updated}

\item{predictions}{The new predcited values for the missing cells}
}
\value{
The updated imputed value table
}
\description{
This function replaces the current imputed values for one covariate with a new set of predictions
}
