% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputationMlHelpers.R
\name{validateLogical}
\alias{validateLogical}
\title{Validate a logical value}
\usage{
validateLogical(x, name)
}
\arguments{
\item{x}{The value to validate}

\item{name}{The argument name used in the error messages}
}
\value{
The validated logical value
}
\description{
This function validates that the input is a single logical value
}
