% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgboostImputer.R
\name{validateXgboostChoice}
\alias{validateXgboostChoice}
\title{Validate an iterative xgboost choice argument}
\usage{
validateXgboostChoice(x, name, choices)
}
\arguments{
\item{x}{The value to validate}

\item{name}{The argument name used in error messages}

\item{choices}{The allowed character values}
}
\value{
The validated character value
}
\description{
This function checks that an iterative xgboost imputation argument is supplied
as a single character value and matches one of the allowed choices.
}
