% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgboostImputer.R
\name{validateXgboostDevice}
\alias{validateXgboostDevice}
\title{Validate a device setting}
\usage{
validateXgboostDevice(device)
}
\arguments{
\item{device}{The device setting to validate}
}
\value{
\code{NULL} or the validated device name string
}
\description{
This helper checks that the iterative xgboost device argument is either
\code{NULL} or a single character value describing a supported compute device
}
